requirements-test.txt
absl-py>=0.7.1
pytest==6.2.5
